third_party/blink/web_tests/flag-specific/disable-layout-ng/fullscreen/full-screen-with-flex-item-expected.txt

Enter Fullscreen
EVENT(webkitfullscreenchange)
EVENT(webkitfullscreenchange)
EXPECTED (584 == '584') OK
END OF TEST
